feat(FX-3443): move saved search helpers to SavedSearch/Utils
