+++ src/v2/Components/ArtworkFilter/SavedSearch/Utils/aggregationForFilter.ts
@@ -1,0 +1,13 @@
+import { Aggregations } from "../../ArtworkFilterContext"
+import { aggregationNameFromFilter } from "./constants"
+
+export const aggregationForFilter = (
+  filterKey: string,
+  aggregations: Aggregations
+) => {
+  const aggregationName = aggregationNameFromFilter[filterKey]
+  const aggregation = aggregations!.find(
+    value => value.slice === aggregationName
+  )
+  return aggregation
+}
