+++ src/v2/Components/ArtworkFilter/SavedSearch/Utils/constants.ts
@@ -1,4 +1,4 @@
-import { Slice } from "../ArtworkFilterContext"
+import { Slice } from "../../ArtworkFilterContext"
 
 export const shouldExtractValueNamesFromAggregation = [
   "locationCities",
