+++ src/v2/Components/ArtworkFilter/Utils/getElementParams.ts
@@ -1,0 +1,5 @@
+export const getElementParams = element => {
+  const height = element.clientHeight
+  const top = element.getBoundingClientRect().top
+  return { height, top }
+}
